Updated files
Changed the name of home back to index. Made the buttons link to different html files. Made the title link to the home page. Made the navigation bar its own html file so it can be loaded into every html page. Made a javascript file to load the navigation bar.